+++ includes/math/spk_vector3.hpp
@@ -85,9 +85,9 @@ namespace spk
 			}
 			else
 			{
-				x = p_object[L"x"].as<int>();
-				y = p_object[L"y"].as<int>();
-				z = p_object[L"z"].as<int>();
+				x = p_object[L"x"].as<long>();
+				y = p_object[L"y"].as<long>();
+				z = p_object[L"z"].as<long>();
 			}
 		}
